+++ src/widgets/auth/components/sign-up-field/SignUpField.tsx
@@ -1,34 +1,38 @@
 import { useForm } from 'react-hook-form'
-import { useAuth } from '@features/auth/hooks'
+import { useAuth } from '@/src/widgets/auth/hooks'
 import { Button, Form } from '@shared/ui'
 import {
   FormField,
   FieldLabel,
   FieldInput,
   FieldErrorMessage,
-} from '@shared/ui/form-field/FormField'
+} from '@/src/shared/ui/form-field/FormField'
 import {
   emailValidation,
+  nicknameValidation,
   passwordValidation,
-} from '@features/auth/lib/form-validation'
-import type { SignInFieldData } from '@features/auth/types/auth.type'
+  passwordConfirmationValidation,
+} from '@/src/widgets/auth/lib/form-validation'
+import type { SignUpFieldData } from '@/src/widgets/auth/types/auth.type'
 
-export default function SignInField() {
+export default function SignUpField() {
   const {
     register,
     handleSubmit,
     formState: { errors },
     setError,
-  } = useForm<SignInFieldData>({
+  } = useForm<SignUpFieldData>({
     mode: 'onBlur',
     defaultValues: {
       email: '',
+      nickname: '',
       password: '',
+      passwordConfirmation: '',
     },
   })
 
-  const { signInSubmit } = useAuth()
-  const onSubmit = signInSubmit(setError)
+  const { signUpSubmit } = useAuth()
+  const onSubmit = signUpSubmit(setError)
 
   return (
     <Form
@@ -41,14 +45,32 @@ export default function SignInField() {
         <FieldErrorMessage />
       </FormField>
 
+      <FormField {...register('nickname', nicknameValidation)} errors={errors}>
+        <FieldLabel>닉네임</FieldLabel>
+        <FieldInput type="text" placeholder="닉네임을 입력해주세요" />
+        <FieldErrorMessage />
+      </FormField>
+
       <FormField {...register('password', passwordValidation)} errors={errors}>
         <FieldLabel>비밀번호</FieldLabel>
         <FieldInput type="password" placeholder="비밀번호를 입력해주세요" />
         <FieldErrorMessage />
       </FormField>
 
+      <FormField
+        {...register('passwordConfirmation', passwordConfirmationValidation)}
+        errors={errors}
+      >
+        <FieldLabel>비밀번호 확인</FieldLabel>
+        <FieldInput
+          type="password"
+          placeholder="비밀번호를 한번 더 입력해주세요"
+        />
+        <FieldErrorMessage />
+      </FormField>
+
       <Button variant="primary" type="submit">
-        로그인
+        회원가입
       </Button>
     </Form>
   )
