+++ src/widgets/auth/lib/form-validation.ts
@@ -1,5 +1,5 @@
 import type { RegisterOptions } from 'react-hook-form'
-import type { SignUpFieldData } from '@features/auth/types/auth.type'
+import type { SignUpFieldData } from '@/src/widgets/auth/types/auth.type'
 
 export const emailValidation = {
   required: '이메일은 필수 입력입니다.',
