+++ src/widgets/auth/lib/getCookies.ts
