+++ src/widgets/auth/types/auth.type.ts
